requirements/requirements.txt
asgiref==3.5.0
Django==3.2.11
django-ckeditor==6.2.0
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.0
Markdown==3.3.6
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
zipp==3.7.0
